Code Refactoring done and Program Enrolment added
