+++ src/models/individualModel.js
@@ -1,13 +1,11 @@
 import {
   Individual,
   ModelGeneral as General,
-  Observation,
-  Concept,
-  ConceptAnswer,
   Gender,
   AddressLevel,
   SubjectType
 } from "openchs-models";
+import { mapObservation } from "./observationModel";
 
 export const mapIndividual = individualDetails => {
   const individual = General.assignFields(
@@ -32,50 +30,6 @@ export const mapIndividual = individualDetails => {
   individual.lowestAddressLevel = addressLevel;
 
   return individual;
-};
-
-
-const mapObservation = (objservationList) => {
-    if (objservationList)
-        return objservationList.map(observation => {
-            return mapConcept(observation);
-        });
-};
-
-const mapConceptAnswer = (conceptAnswer) => {
-  if (conceptAnswer) {
-      let mapConceptAnswer = {};
-      for (let [key, value] of Object.entries(conceptAnswer)) {
-          let conceptAnswer = new ConceptAnswer();
-          conceptAnswer = mapAnswerConceptObs(value);
-          mapConceptAnswer[key] = conceptAnswer;
-      }
-      return mapConceptAnswer;
-  }
-};
-
-const mapAnswerConceptObs = (observationJson) => {
-    if (observationJson) {
-        const observation = new Observation();
-        const concept = General.assignFields(observationJson, new Concept(), ["uuid", "name"]);
-        concept.datatype = observationJson["dataType"];
-          observation.concept = concept;
-          return observation;  
-    }
-};
-
-const mapConcept = (observationJson) => {
-    if (observationJson) {
-        const observation = new Observation();
-        const concept = General.assignFields(observationJson.concept, new Concept(), ["uuid", "name"]);
-        concept.datatype = observationJson.concept["dataType"];
-        if (observationJson.concept["answers"] != undefined) {
-          concept.answers = mapConceptAnswer(observationJson.concept["answers"]);
-        }
-        observation.concept = concept;
-        observation.valueJSON = JSON.stringify(concept.getValueWrapperFor(observationJson.value));
-        return observation;  
-    }
 };
 
 //subject Dashboard profile Tab
