+++ src/ruleEvaluation/decisionRule.js
@@ -1,0 +1,17 @@
+import * as rulesConfig from 'rules-config';
+import lodash from 'lodash';
+import moment from 'moment';
+
+export const decisionRule = async (rule,entity) => {
+    const defaultDecisions = {
+        "enrolmentDecisions": [],
+        "encounterDecisions": [],
+        "registrationDecisions": []
+    };
+    const ruleFunc = eval(rule);
+    const ruleDecisions = ruleFunc({
+        params: {decisions: defaultDecisions, entity},
+        imports: {rulesConfig, lodash, moment}
+    });
+    return ruleDecisions;
+}
